Model-Implementation/requirements.txt
ipykernel
numpy
matplotlib
tensorflow
IPython
scikit-learn
seaborn
fastapi
uvicorn
python-multipart
pillow
opencv-python